pyproject.toml
[project]
name = "etl-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9, <3.13"
dependencies = [
    "apache-airflow>=3.0.6",
    "apache-airflow-providers-amazon>=9.9.0",
    "apache-airflow-providers-fab>=2.2.1",
    "apache-airflow-providers-mysql>=6.3.1",
    "apache-airflow-providers-postgres>=6.2.0",
    "ipykernel>=6.30.1",
    "mysqlclient>=2.2.7",
    "numpy>=2.0.2",
    "pandas>=2.3.3",
    "psycopg2-binary>=2.9.10",
]

[tool.uv.workspace]
members = [
    "airflow-project",
]
